# Fixed Redundant Difficulty Calculations