Fixed multiple calculations where difficulty had already been calculated.

- Added Variable for Difficulty Value.

- Added a check to verify if DifficultyValue has already been calculated; if not, it calculates and stores the value to prevent redundant calculations.

- If CountTopWeightedStrains() or RelevantNoteCount() is called, it checks if they have already been calculated; if not, it recalculates and stores the value.