@stanriders 2025-03-23
Mind doing a benchmark? I have a feeling that the difference is insignificant. Another thing I'm not sure about is if it doesn't break the pp counter in the client