@TheDark98 2025-03-23
> Mind doing a benchmark? I have a feeling that the difference is insignificant. Another thing I'm not sure about is if it doesn't break the pp counter in the client
### Latest tests:
**Current osu-dev:**
![standard](https://github.com/user-attachments/assets/6352651b-a72e-4ff1-8c8a-6e52ea887f5d)
**With My Optimization:**
![new](https://github.com/user-attachments/assets/2bef6271-12fa-4439-86cd-5a2f3945245a)

I've done some benchmarks, and on average, when recalculating only the Dark98 OSU profile, it's 12.7% faster than the current one.
Also, it doesn't break anything because I only stored a value; nothing else changed.