@stanriders 2025-03-23
PP counter also doesn't seem to work

In-game:
![image](https://github.com/user-attachments/assets/05d35c9a-ebf0-48f9-a8b9-3837b2ca1414)
Result screen:
![image](https://github.com/user-attachments/assets/b90d6316-3f26-4b9b-87f8-219e283b45ac)