@TheDark98 2025-03-23
> Test failures seem relevant

This is not my fault; it matches the PP-Dev tests.
![image](https://github.com/user-attachments/assets/a702b40d-fd5d-41a4-8e11-56f526501a38)
Also, by looking at what is failing, you can clearly see where it is failing.




> PP counter also doesn't seem to work
> 
> In-game: ![image](https://private-user-images.githubusercontent.com/8269193/425856095-05d35c9a-ebf0-48f9-a8b9-3837b2ca1414.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NDI3NTY0MzAsIm5iZiI6MTc0Mjc1NjEzMCwicGF0aCI6Ii84MjY5MTkzLzQyNTg1NjA5NS0wNWQzNWM5YS1lYmYwLTQ4ZjktYThiOS0zODM3YjJjYTE0MTQucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MDMyMyUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTAzMjNUMTg1NTMwWiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9MDIzNTBkZTI0MmFjZWRjYTRhMTI3NzRhYjlhZDZmMmQ2ZjBlMjg4Mzk3ZDU4ZjNlNTc5NWM4ODFkOTRjN2ExZiZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.155pAbZ5upG4w5AdOIbUwv2DcwgA9hwbaIwthyC2K7c) Result screen: ![image](https://private-user-images.githubusercontent.com/8269193/425856115-b90d6316-3f26-4b9b-87f8-219e283b45ac.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NDI3NTY0MzAsIm5iZiI6MTc0Mjc1NjEzMCwicGF0aCI6Ii84MjY5MTkzLzQyNTg1NjExNS1iOTBkNjMxNi0zZjI2LTRiOWItODdmOC0yMTllMjgzYjQ1YWMucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MDMyMyUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTAzMjNUMTg1NTMwWiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9ZjE0Yjc2MmZhYjczOGVjOGJlNWRmZWJiMTMyYzY1MTU4MTNkZDU3MGIwZjg3OTVhYzY2YmU0ZTMwZGFjN2ZiMCZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.NaQ9IizBl18KBOMXSEXTF80g2lw18ZtKC3Cj2y06Uqo)

Thanks for the feedback! I'll check why it breaks the PP counter!