@stanriders 2025-03-23
Running proper benchmarks doesn't seem to yield any meaningful difference. 

Benchmarks: https://github.com/ppy/osu/pull/32528/commits/10fd38ac146e83202c846d40b356af432c69121f
I'm using the same map for all rulesets, so for all non-osu rulesets its a converted map but I don't think that should make any difference.

**Before**:
| Method                   | Mean      | Error    | StdDev   | Gen0      | Gen1      | Gen2     | Allocated |
|------------------------- |----------:|---------:|---------:|----------:|----------:|---------:|----------:|
| CalculateDifficultyOsu   |  42.36 ms | 0.486 ms | 0.380 ms |  916.6667 |  750.0000 |        - |   46.3 MB |
| CalculateDifficultyTaiko |  55.80 ms | 1.105 ms | 1.316 ms |  888.8889 |  777.7778 | 222.2222 |  36.31 MB |
| CalculateDifficultyCatch | 103.20 ms | 1.245 ms | 1.165 ms | 1666.6667 | 1500.0000 | 500.0000 |  62.61 MB |
| CalculateDifficultyMania |  39.78 ms | 0.668 ms | 1.170 ms |  846.1538 |  769.2308 | 230.7692 |  31.64 MB |

**After**:
| Method                   | Mean      | Error    | StdDev   | Gen0      | Gen1      | Gen2     | Allocated |
|------------------------- |----------:|---------:|---------:|----------:|----------:|---------:|----------:|
| CalculateDifficultyOsu   |  43.36 ms | 0.659 ms | 0.515 ms |  916.6667 |  750.0000 |        - |  45.71 MB |
| CalculateDifficultyTaiko |  53.59 ms | 1.063 ms | 1.943 ms |  888.8889 |  777.7778 | 222.2222 |  35.85 MB |
| CalculateDifficultyCatch | 103.89 ms | 1.787 ms | 1.584 ms | 1666.6667 | 1500.0000 | 500.0000 |  62.61 MB |
| CalculateDifficultyMania |  39.74 ms | 0.553 ms | 0.517 ms |  846.1538 |  769.2308 | 230.7692 |  31.64 MB |