@stanriders 2025-03-23
Running `diffcalc.Calculate()` 10 times in a row doesn't yield any difference either.

**Before**:
| Method                   | Mean      | Error     | StdDev    | Gen0       | Gen1       | Gen2      | Allocated |
|------------------------- |----------:|----------:|----------:|-----------:|-----------:|----------:|----------:|
| CalculateDifficultyOsu   | 737.35 ms | 14.510 ms | 21.268 ms | 12000.0000 | 11000.0000 | 3000.0000 | 462.98 MB |

**After**:
| Method                   | Mean      | Error     | StdDev    | Gen0       | Gen1       | Gen2      | Allocated |
|------------------------- |----------:|----------:|----------:|-----------:|-----------:|----------:|----------:|
| CalculateDifficultyOsu   | 748.38 ms | 14.555 ms | 23.914 ms | 12000.0000 | 11000.0000 | 3000.0000 | 457.14 MB |