@TheDark98 2025-03-23
> Sorry, but this PR does nothing

After seeing the benchmarks outside my machine, I agree.